# docs/api/swagger.yaml
openapi: 3.0.0
info:
  title: Tourist Tour Guides API
  description: API documentation for the Tourist Tour Guides application.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server
paths:
  /auth/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
  /auth/register:
    post:
      summary: User registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                email:
                  type: string
      responses:
        '201':
          description: Successful registration
...